Fxi build due to https://github.com/ClickHouse/ClickHouse/pull/64108 which move StringUtils.h from src/Common/StringUtils/StringUtils.h to src/Common/StringUtils.h.
To avoid conflict, we rename StringUtils.h to GlutenStringUtils.h

(cherry picked from commit 60f2b9e16accafd97e3135feb1036d877dbb363e)
