+++ cpp-ch/local-engine/tests/gtest_utils.cpp
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 #include <gtest/gtest.h>
-#include <Common/StringUtils.h>
+#include <Common/GlutenStringUtils.h>
 
 using namespace local_engine;
 
 TEST(TestStringUtils, TestExtractPartitionValues)
 {
     std::string path = "/tmp/col1=1/col2=test/a.parquet";
-    auto values = StringUtils::parsePartitionTablePath(path);
+    auto values = GlutenStringUtils::parsePartitionTablePath(path);
     ASSERT_EQ(2, values.size());
     ASSERT_EQ("col1", values[0].first);
     ASSERT_EQ("1", values[0].second);
